chore: moved session things around to help readability for more major changes
